requirements.txt
py==1.8.1
python-dateutil==2.8.1
python-crontab==2.5.1
pyppeteer==0.2.2
pdf2image==1.14.0
SQLAlchemy==1.3.16
psycopg2==2.8.5
PyJWT==1.7.1
